Use the same component for speaker grouping chips everywhere
